inticure_backend/requirements.txt
Django==4.1
django-cors-headers==3.13.0
django-cron==0.6.0
cryptography>=42.0.1
djangorestframework==3.13.1
google-auth>=2.15.0
google-auth-oauthlib==1.0.0
mysqlclient==2.1.1
oauthlib==3.2.2
Pillow==9.2.0
requests==2.28.1
sqlparse==0.4.2
stripe==5.0.0
twilio==7.15.4


